feat(config): make metatables play well with default configs
Previously, if one of the values in configuration table passed to
the `setup` function was a table that had a metatable with an
`__index` field, and that configuration value is also merged with a
default value, the metatable would be ignored, because when the
tables were merged only the keys actually present on the
user-provided table were considered. With this patch, the merged
configuration looks up each key dynamically, so that metatables on
configuration values work as expected.